Moved profile initialization to registration
